+++ NEWS.md
@@ -1,7 +1,9 @@
 # landscapemetrics 1.4.5
 * Improvements
     * `lsm_p_gyrate` has an argument to force the cell centroid to be within patch 
-
+* New gunctions
+    * `get_centroid()` returns the coordinates of each patch centroid
+    
 # landscapemetrics 1.4.4
 * Improvements
     * Set labels = FALSE as default for all plotting functions (messy for larger raster)